requirements.txt
Django==4.1.5
envparse==0.2.0
python-decouple==3.7
psycopg2==2.9.5
uvicorn[standard]
gunicorn==20.1.0
factory-boy==3.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.14.0
django-debug-toolbar
django-softdelete==0.10.5
# pycodestyle==2.11.0
# pylint==2.17.5
flake8==6.1.0